Increased parallelism!
Solved a problem that bugged me since the beginning: how to make latency
for the server as low as possioble and maximize CPU core usage.

We use a ThreadPool which gets invoked via generic_do_async() in
Servers.cpp for the more expensive rpc_* calls.  All the ones that may
hit the db and take a while, or may churn on the CPU (such as the
merkle-related ones) immediately go async and schedule their work on the
threadpool.

The rule will be only the most trivial rpc methods generate results
immediately, the rest schedule work via the threadpool.

This will help us scale and leverage as many cores as possible.  It also
allows us to have many many clients "living" in the server's thread
(thus reducing lock usage).

I'm excited about this design.

+++ Controller.cpp
@@ -839,7 +839,7 @@ auto Controller::stats() const -> Stats
     st["Storage"] = storage->statsSafe();
     QVariantMap misc;
     misc["extant thread pool jobs"] = Util::ThreadPool::ExtantJobs();
-    misc["thread pool jobs (lifetime)"] = Util::ThreadPool::NumJobsSubmitted();
+    misc["thread pool jobs (lifetime)"] = qulonglong(Util::ThreadPool::NumJobsSubmitted());
     st["Misc"] = misc;
     return st;
 }
